Finalização das classes sobre copetencia, adicionamento de novos métodos get, requisições http do tipo get testadas, conexão com o bnco de dados remoto

+++ BackendContratanti/src/main/java/com/projeto/BackendContratanti/Dto/CompetenciasRequestDTO.java
@@ -1,4 +1,6 @@
 package com.projeto.BackendContratanti.Dto;
 
-public record CompetenciasRequestDTO (Integer id, String nome, String email){
+import java.math.BigInteger;
+
+public record CompetenciasRequestDTO (BigInteger id, String nome, String email){
 }
