+++ BackendContratanti/src/main/java/com/projeto/BackendContratanti/Dto/CompetenciasResponseDTO.java
@@ -2,7 +2,9 @@ package com.projeto.BackendContratanti.Dto;
 
 import com.projeto.BackendContratanti.Model.Competencias;
 
-public record CompetenciasResponseDTO (Integer id, String nome, String email){
+import java.math.BigInteger;
+
+public record CompetenciasResponseDTO (BigInteger id, String nome, String email){
     public CompetenciasResponseDTO(Competencias competencias){
         this(competencias.getId(), competencias.getNome(), competencias.getEmail());
     }
